tree node model and methods for accessing node information
