+++ src/pages/Feeds/components/NodeDetails.tsx
@@ -2,8 +2,10 @@ import React from "react";
 import { IPluginItem } from "../../../api/models/pluginInstance.model";
 import { Button, Grid, GridItem } from "@patternfly/react-core";
 import { ShareAltIcon, InfrastructureIcon } from "@patternfly/react-icons";
+import TreeNodeModel, { INode } from "../../../api/models/tree-node.model";
 interface INodeProps {
   selected: IPluginItem;
+  items: IPluginItem[];
 }
 
 class NodeDetails extends React.Component<INodeProps> {
@@ -17,8 +19,15 @@ class NodeDetails extends React.Component<INodeProps> {
     // Stub - To be done
   }
 
+  // Description: root node or leaf nodes in the graph will not have the 'share this pipeline' button
+  isNodePipelineRoot(item: IPluginItem ) {
+    const { items } = this.props;
+    // Find out from items if this node is a leaf or root node
+    return (!TreeNodeModel.isRootNode(item) && !TreeNodeModel.isLeafNode(item, items));
+  }
+
   render() {
-    const { selected } = this.props;
+    const { selected, items } = this.props;
     return (
       <React.Fragment>
         <div>
@@ -29,16 +38,22 @@ class NodeDetails extends React.Component<INodeProps> {
             Pipeline chart
           </GridItem>
           <GridItem className="pf-u-p-sm" sm={12} md={6}>
-              <label>From this node:</label>
-              <div className="btn-div">
-                <Button variant="tertiary" isBlock  onClick={this.handleAddNewNode}>
-                  <InfrastructureIcon /> Add new node(s)...
-                </Button>
-                <Button variant="tertiary" isBlock onClick={this.handleSharePipeline}>
-                  <ShareAltIcon /> Share this pipeline...
-                </Button>
-              </div>
-
+            <label>From this node:</label>
+            <div className="btn-div">
+              <Button
+                variant="tertiary"
+                isBlock
+                onClick={this.handleAddNewNode} >
+                <InfrastructureIcon /> Add new node(s)...
+              </Button>
+              {
+                this.isNodePipelineRoot(selected) && (
+                  <Button variant="tertiary" isBlock onClick={this.handleSharePipeline}>
+                    <ShareAltIcon /> Share this pipeline...
+                  </Button>
+                )
+              }
+            </div>
           </GridItem>
         </Grid>
       </React.Fragment>
